Tidy Results from a List of Estimators

This function applies extracts a "tidy" output from a fitted model, or from a named list of fitted models. The information is extracted using the `hyotheses()` function from the `marginaleffects` package.

Usage

extract_estimates(..., vcov = NULL, alpha = 0.05, null = 0, term = NULL)

Arguments

...

A named list of fitted model objects. Each model is passed to `marginaleffects::hypotheses()`.

vcov

An optional variance-covariance matrix or a function to compute it. Passed to `extract_estimates`.

alpha

A numeric value between 0 and 1 specifying the significance level for confidence intervals (default is 0.05).

null

A numeric value representing the null hypothesis for the parameter estimates (default is 0).

term

An optional character vector of terms to filter the output. If specified, only rows with matching terms are returned.

Value

A data frame combining the tidied results from all models in the list. Columns include:

estimator

The label for the estimator (corresponding to the list names).

term

The name of the parameter (e.g., model term).

estimate

The estimated value of the parameter.

p.value

The p-value for the hypothesis test of the parameter.

conf.low

The lower bound of the confidence interval.

conf.high

The upper bound of the confidence interval.

Examples

library(SimpleDesign)
extract_estimates(
  `I` = lm(mpg ~ wt, data = mtcars),
  `II` = lm(mpg ~ wt + disp, data = mtcars),
  alpha = 0.01,
  term = "wt"
)
#>   estimator term  estimate      p.value  conf.low  conf.high
#> 1         I   wt -5.344472 1.188493e-21 -6.784620 -3.9043227
#> 2        II   wt -3.350825 3.996990e-03 -6.349421 -0.3522301